Move corresponding tests.
